lovelace-server/requirements.txt
APScheduler==3.9.1.post1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==22.1.0
bcrypt==4.0.1
bidict==0.22.0
blinker==1.5
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
colorhash==1.2.1
commonmark==0.9.1
configparser==5.3.0
cryptography==38.0.4
Deprecated==1.2.13
dnspython==2.2.1
Flask==2.2.2
flask-expects-json==1.7.0
Flask-HTTPAuth==4.7.0
Flask-Limiter==2.8.1
Flask-Mail==0.9.1
Flask-MonitoringDashboard==3.1.1
Flask-PyMongo==2.3.0
Flask-SocketIO==5.3.2
future==0.18.2
greenlet==2.0.1
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
limits==2.7.2
MarkupSafe==2.1.1
numpy==1.24.1
packaging==22.0
prometheus-client==0.16.0
prometheus-flask-exporter==0.21.0
psutil==5.9.4
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pymongo==4.3.3
pyotp==2.8.0
pyrsistent==0.19.2
python-dotenv==0.21.0
python-engineio==4.3.4
python-json-logger==2.0.4
python-socketio==5.7.2
pytz==2022.7.1
requests==2.28.1
rich==12.6.0
scipy==1.10.0
simple-websocket==0.9.0
six==1.16.0
SQLAlchemy==1.4.46
typing_extensions==4.4.0
tzlocal==2.0.0
urllib3==1.26.13
validate-email==1.3
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
WTForms==3.0.1
